fixes problems with nth and the debug function, while working on thiago's branch
