+++ webui/react/package.json
@@ -134,7 +134,7 @@
     "jsdom": "^16.7.0",
     "morgan": "^1.10.0",
     "npm-force-resolutions": "0.0.10",
-    "playwright-page-model-base": "npm:@hpe.com/playwright-page-model-base@^0.2.5",
+    "playwright-page-model-base": "npm:@hpe.com/playwright-page-model-base@^0.3.0",
     "prettier": "^3.2.5",
     "request": "^2.88.2",
     "resize-observer-polyfill": "^1.5.1",
