+++ webui/react/src/e2e/tests/experimentList.spec.ts
@@ -286,8 +286,13 @@ test.describe('Experiment List', () => {
     );
   });
 
-  test('Multi-sort menu', async ({ dev }) => {
-    const checkTableOrder = async (firstKey: keyof ExperimentBase, secondKey: keyof ExperimentBase) => {
+  test('Multi-sort menu', async () => {
+    const multiSortMenu = projectDetailsPage.f_experimentList.tableActionBar.multiSortMenu;
+    const secondRow = multiSortMenu.multiSort.rows.nth(1);
+    const checkTableOrder = async (
+      firstKey: keyof ExperimentBase,
+      secondKey: keyof ExperimentBase,
+    ) => {
       const experimentList: ExperimentBase[] = JSON.parse(await detExecSync('experiment ls'));
 
       return [
@@ -298,7 +303,6 @@ test.describe('Experiment List', () => {
         },
       ];
     };
-    const multiSortMenu = projectDetailsPage.f_experimentList.tableActionBar.multiSortMenu;
 
     const sortingScenario = async (
       firstSortBy: string,
@@ -318,9 +322,7 @@ test.describe('Experiment List', () => {
         await firstRow.order.selectMenuOption(firstSortOrder);
 
         await multiSortMenu.multiSort.add.pwLocator.click();
-        
-        const secondRow = multiSortMenu.multiSort.rows.nth(1);
-        dev.debugComponentVisible(secondRow);
+
         await secondRow.column.selectMenuOption(secondSortBy);
         await secondRow.order.selectMenuOption(secondSortOrder);
 
@@ -330,15 +332,15 @@ test.describe('Experiment List', () => {
       });
     };
 
-    await sortingScenario('ID', '9 → 0', 'Start time', 'Oldest → Newest',  async () => {
+    await sortingScenario('ID', '9 → 0', 'Start time', 'Oldest → Newest', async () => {
       const [higher, lower] = await checkTableOrder('id', 'startTime');
-      await expect(higher.id).toBeGreaterThan(lower.id as number);
-      },);
+      expect(higher.id).toBeGreaterThan(lower.id as number);
+    });
 
-    await sortingScenario('Trial count', '0 → 9', 'Searcher', 'A → Z',  async () => {
+    await sortingScenario('Trial count', '0 → 9', 'Searcher', 'A → Z', async () => {
       const [first, last] = await checkTableOrder('numTrials', 'searcherType');
-      await expect(first.numTrials).toBeLessThanOrEqual(last.numTrials as number);
-      },);
+      expect(first.numTrials).toBeLessThanOrEqual(last.numTrials as number);
+    });
   });
 
   test('Datagrid Functionality Validations', async ({ authedPage }) => {
